Icloud_routine/requirements.txt
pandas
openpyxl
python-dotenv
requests
aiogram==3.13.0
asyncio